done with the event timeline layout
