+++ src/pages/event/event.data.tsx
@@ -1,3 +1,10 @@
+interface IEventData {
+  id: number
+  location: string
+  title: string
+  date: string
+}
+
 export const EventInterestData = [
   'Blockchain',
   'crypto',
@@ -14,4 +21,55 @@ export const EventInterestData = [
   'Artist',
   'Non-fungible',
   'Solana',
+]
+
+export const eventMockData: IEventData[] = [
+  {
+    id: 1,
+    location: 'Le Carrousel du Louvre (Paris, France)',
+    title: '5th Edition Paris Blockchain Week',
+    date: 'April 8-12, 2024',
+  },
+  {
+    id: 2,
+    location: 'Rai, Amsterdam.',
+    title: 'Blockchain Expo Europe',
+    date: 'October 1-02, 2024',
+  },
+  {
+    id: 3,
+    location: 'AsiaWorld Expo, Airport Expo Blvd...',
+    title: 'Wow summit in Hong Kong',
+    date: 'march, 26-27,2024',
+  },
+  {
+    id: 4,
+    location: 'Dubai, Festival Arena',
+    title: 'Blockchain Life 2024',
+    date: 'April 15-16, 2024',
+  },
+  {
+    id: 5,
+    location: 'Le Carrousel du Louvre (Paris, France)',
+    title: '5th Edition Paris Blockchain Week',
+    date: 'April 8-12, 2024',
+  },
+  {
+    id: 6,
+    location: 'Rai, Amsterdam.',
+    title: 'Blockchain Expo Europe',
+    date: 'October 1-02, 2024',
+  },
+  {
+    id: 7,
+    location: 'AsiaWorld Expo, Airport Expo Blvd...',
+    title: 'Wow summit in Hong Kong',
+    date: 'march, 26-27,2024',
+  },
+  {
+    id: 8,
+    location: 'Dubai, Festival Arena',
+    title: 'Blockchain Life 2024',
+    date: 'April 15-16, 2024',
+  },
 ]
